Get the number of reservation in the bell
