+++ front-end/src/views/MaterialsView.vue
@@ -28,7 +28,7 @@
 						<div
 							class="absolute inline-flex items-center justify-center w-6 h-6 text-xs font-bold text-white bg-red-500 border-2 border-white rounded-full -top-2 -right-2 dark:border-gray-900"
 						>
-							20
+							{{ reservation.length }}
 						</div>
 					</IconButton>
 				</div>
@@ -39,7 +39,7 @@
 				</template>
 			</Modal>
 			<div v-if="notificationCenter">
-				<ApprouvalCenter />
+				<ApprouvalCenter :users="users" :reservation="reservation" />
 			</div>
 			<!-- <h2 class="text-2xl font-bold dark:text-dark-font">List of all materials</h2> -->
 			<div class="cards mt-5" v-if="materialsComponent">
@@ -53,7 +53,7 @@
 </template>
 
 <script lang="ts" setup>
-import { computed, ref } from 'vue';
+import { computed, ref, onMounted } from 'vue';
 import Modal from '@/components/common/Modal.vue';
 import ListMaterials from '@/components/materials/ListMaterials.vue';
 import ButtonsMaterials from '@/components/materials/ButtonsMaterials.vue';
@@ -71,6 +71,7 @@ import { Roles } from '@/store/interfaces/auth.interfaces';
 import { useAuthStore } from '@/store/modules/auth.store';
 import Bell from '@/components/common/svg/Bell.vue';
 import ApprouvalCenter from '@/components/materials/ApprouvalCenter.vue';
+import { useMaterialStore } from '@/store/modules/material.store';
 
 const showModal = ref(false);
 const notificationCenter = ref(false);
@@ -79,6 +80,9 @@ const materialsComponent = ref(true);
 const authStore = useAuthStore();
 const user = computed(() => authStore.user);
 const userRole = computed(() => user.value?.role);
+const materialStore = useMaterialStore();
+const reservation = computed(() => materialStore.pendingMaterials);
+const users = ref([]);
 
 pdfMake.vfs = pdfFonts.pdfMake.vfs;
 let base64Image = null;
@@ -95,6 +99,16 @@ fetch(CodingToolsLogo)
 	.catch((error) => {
 		console.error("Erreur lors du chargement de l'image :", error);
 	});
+
+onMounted(() => {
+	materialStore.getPendingMaterials().then(() => {
+		reservation.value.forEach((res1) => {
+			materialStore.getUserById(res1.borrowingHistory.borrowingUser).then((res) => {
+				users.value.push(res);
+			});
+		});
+	});
+});
 
 const showGraph = () => {
 	graphComponent.value = !graphComponent.value;
